Checkpoint execution of requests, bookmarks and process lists
